+++ vite.config.ts
@@ -1,5 +1,6 @@
 import tailwindcss from '@tailwindcss/vite';
 import react from '@vitejs/plugin-react';
+import viteCompression from 'vite-plugin-compression';
 import packageJson from './package.json';
 import path from 'path';
 import {defineConfig} from 'vite';
@@ -8,7 +9,7 @@ export default defineConfig({
   define: {
     __APP_VERSION__: JSON.stringify(packageJson.version),
   },
-  plugins: [react(), tailwindcss()],
+  plugins: [react(), tailwindcss(), viteCompression({ algorithm: 'brotliCompress', ext: '.br' })],
   resolve: {
     alias: {
       '@': path.resolve(__dirname, '.'),